+++ js/form.js
@@ -1,10 +1,28 @@
 /*! js/form.js | MIT License | github.com/visicode/Balzac.js */
 'use strict';
-import './document.js';
 
 const USER_INVALID_CLASS = 'user-invalid';
 
+function setFocus() {
+	if (document.activeElement === document.body) // no element already focused
+		Array.from(document.forms).find(form => {
+			const rect = form.getBoundingClientRect();
+			if (rect.top < window.innerHeight && rect.bottom >= 0) { // form visible
+				const first = Array.from(form.elements).find(element =>
+					element.tabIndex > -1
+					&& !element.disabled
+					&& !element.readOnly);
+				if (first) {
+					first.focus();
+					return true; // input found, exit forms loop
+				}
+			}
+			return false;
+		});
+}
+
 /**
+ * - Automatic focus on the first input field of the currently visible form.
  * - Adds "user-invalid" class to invalid form and form elements after user interaction, emulating the poorly-supported ":user-invalid" pseudo-class.
  * - Clears content of HTML output elements when resetting form.
  */
@@ -23,6 +41,7 @@ function fixForm(form) {
 			});
 		}
 	});
+
 	form.addEventListener('reset', _ => {
 		Array.from(form.elements).forEach(element => {
 			if (element.validity) {
@@ -32,7 +51,11 @@ function fixForm(form) {
 			}
 		});
 		form.classList.remove(USER_INVALID_CLASS);
+		setFocus();
 	});
+
+	window.addEventListener('scroll', setFocus);
+	setFocus();
 }
 
 export default {