+++ js/math.js
@@ -3,8 +3,9 @@
 
 /**
  * Returns a numeric expression rounded to a specified number of fractional digits.
- * @param x The value to be rounded.
- * @param decimals The number of fractional digits.
+ * @param {number} x The value to be rounded.
+ * @param {number} decimals The number of fractional digits.
+ * @returns {number} The value rounded to the specified number of fractional digits.
  */
 Math.roundTo || Object.defineProperties(Math, {
 	roundTo: {