+++ js/number.js
@@ -3,8 +3,9 @@
 
 /**
  * Checks if a number is between two values, bounds included.
- * @param min The minimum value.
- * @param max The maximum value.
+ * @param {number} min The minimum value.
+ * @param {number} max The maximum value.
+ * @returns {boolean} true if the number is between the specified values, otherwise false.
  */
 Number.prototype.isBetween || Object.defineProperties(Number.prototype, {
 	isBetween: {
@@ -17,8 +18,9 @@ Number.prototype.isBetween || Object.defineProperties(Number.prototype, {
 
 /**
  * Checks if a number is between two values, bounds excluded.
- * @param min The minimum value.
- * @param max The maximum value.
+ * @param {number} min The minimum value.
+ * @param {number} max The maximum value.
+ * @returns {boolean} true if the number is between the specified values, otherwise false.
  */
 Number.prototype.isBetweenExclusive || Object.defineProperties(Number.prototype, {
 	isBetweenExclusive: {